Move auto-registration logic inside the converter
