+++ crates/extensions/c8y_mapper_ext/src/converter.rs
@@ -53,7 +53,9 @@ use std::path::Path;
 use std::path::PathBuf;
 use tedge_actors::LoggingSender;
 use tedge_actors::Sender;
+use tedge_api::entity::EntityTopic;
 use tedge_api::entity_store;
+use tedge_api::entity_store::EntityRegistrationMessage;
 use tedge_api::event::error::ThinEdgeJsonDeserializerError;
 use tedge_api::event::ThinEdgeEvent;
 use tedge_api::topic::get_child_id_from_measurement_topic;
@@ -82,6 +84,7 @@ use tracing::debug;
 use tracing::info;
 use tracing::log::error;
 
+const MQTT_ROOT: &str = "te";
 const C8Y_CLOUD: &str = "c8y";
 const INVENTORY_FRAGMENTS_FILE_LOCATION: &str = "device/inventory.json";
 const SUPPORTED_OPERATIONS_DIRECTORY: &str = "operations";
@@ -241,6 +244,64 @@ impl CumulocityConverter {
             entity_store,
         })
     }
+
+    /// Performs auto-registration process for an entity under a given
+    /// identifier.
+    ///
+    /// If an entity is a service, its device is also auto-registered if it's
+    /// not already registered.
+    ///
+    /// It returns MQTT register messages for the given entities to be published
+    /// by the mapper, so other components can also be aware of a new device
+    /// being registered.
+    fn auto_register_entity(
+        &mut self,
+        entity_id: &str,
+    ) -> Result<Vec<Message>, entity_store::Error> {
+        let mut register_messages = vec![];
+        let (device_id, service_id) = match entity_id.split('/').collect::<Vec<&str>>()[..] {
+            ["device", device_id, "service", service_id, ..] => (device_id, Some(service_id)),
+            ["device", device_id, "", ""] => (device_id, None),
+            _ => return Ok(register_messages),
+        };
+
+        // register device if not registered
+        let device_topic = format!("device/{device_id}//");
+        if self.entity_store.get(&device_topic).is_none() {
+            let device_register_payload = r#"{ "@type": "child-device" }"#.to_string();
+            let device_register_message =
+                Message::new(&Topic::new(&device_topic).unwrap(), device_register_payload)
+                    .with_retain();
+            register_messages.push(device_register_message.clone());
+            self.entity_store
+                .update(EntityRegistrationMessage::try_from(&device_register_message).unwrap())?;
+        }
+
+        // register service itself
+        if let Some(service_id) = service_id {
+            let service_topic = format!("{MQTT_ROOT}/device/{device_id}/service/{service_id}");
+            let service_register_payload = r#"{"@type": "service", "type": "systemd"}"#.to_string();
+            let service_register_message = Message::new(
+                &Topic::new(&service_topic).unwrap(),
+                service_register_payload,
+            )
+            .with_retain();
+            register_messages.push(service_register_message.clone());
+            self.entity_store
+                .update(EntityRegistrationMessage::try_from(&service_register_message).unwrap())?;
+        }
+
+        Ok(register_messages)
+    }
+
+    /// Registers the entity under a given MQTT topic.
+    ///
+    /// If a given entity was registered previously, the function will do
+    /// nothing. Otherwise it will save registration data to memory, free to be
+    /// queried by other components.
+    // fn register_entity(&mut self, topic: String, payload: String) {
+    //     self.entity_store.entry(&topic).or_insert(payload);
+    // }
 
     fn try_convert_measurement(
         &mut self,
@@ -690,7 +751,29 @@ impl Converter for CumulocityConverter {
         &self.mapper_config
     }
     async fn try_convert(&mut self, message: &Message) -> Result<Vec<Message>, ConversionError> {
-        match &message.topic {
+        let mut registration_messages = vec![];
+
+        if let Ok(entity_topic) = EntityTopic::try_from(&message.topic) {
+            if let Ok(register_message) = EntityRegistrationMessage::try_from(message) {
+                if let Err(e) = self.entity_store.update(register_message) {
+                    error!("Could not update device registration: {e}");
+                }
+            } else {
+                // if device is unregistered register using auto-registration
+                if self.entity_store.get(entity_topic.entity_id()).is_none() {
+                    registration_messages =
+                        match self.auto_register_entity(entity_topic.entity_id()) {
+                            Ok(register_messages) => register_messages,
+                            Err(e) => {
+                                error!("Could not update device registration: {e}");
+                                vec![]
+                            }
+                        };
+                }
+            }
+        }
+
+        let mut messages = match &message.topic {
             topic if topic.name.starts_with("tedge/measurements") => {
                 self.size_threshold.validate(message)?;
                 self.try_convert_measurement(message)
@@ -734,7 +817,10 @@ impl Converter for CumulocityConverter {
                     message.topic.name.clone(),
                 )),
             },
-        }
+        }?;
+
+        registration_messages.append(&mut messages);
+        Ok(registration_messages)
     }
 
     fn try_init_messages(&mut self) -> Result<Vec<Message>, ConversionError> {
